feat(home): "Recently translated" slider after collections
Adds a homepage band, right after the collections grid, that reuses the
Mycology "First translations" slider and auto-fills it with the 15 works
most recently brought into a modern translation site-wide.

- Extract the mycology slider into a shared src/components/BookSlider.tsx;
  MycoSlider becomes a thin re-export (single source of truth, mycology
  page unchanged).
- getRecentlyTranslated() in home-data.ts reads the Supabase books_catalog
  mirror (sort=last_translated, indexed) — ~0.5s vs a 15s unindexed Mongo
  full scan for the same 15 rows. Filters visible + hasTranslation, which
  also excludes artworks.
- Section hides itself when the query returns nothing; EN + ES strings added.
