+++ src/lib/home-i18n.ts
@@ -60,6 +60,10 @@ export interface HomeStrings {
   curatedExhibitions: string;
   allCollections: string;
 
+  // Recently translated slider
+  recentlyTranslatedHeading: string;
+  recentlyTranslatedSubtitle: string;
+
   // Discover section
   discoverHeading: string;
   discoverSubtitle: string;
@@ -160,6 +164,8 @@ const en: HomeStrings = {
   curatedExhibitions: 'Browse curated exhibitions',
   allCollections: 'All collections',
 
+  recentlyTranslatedHeading: 'Recently translated',
+  recentlyTranslatedSubtitle: 'The latest works Source Library has brought into a modern, readable translation.',
   discoverHeading: 'Discover',
   discoverSubtitle: 'Translated primary sources from the collection.',
   discoverEmpty: 'Browse the collection to discover translated primary sources.',
@@ -261,6 +267,8 @@ const es: HomeStrings = {
   curatedExhibitions: 'Explorar exposiciones comisariadas',
   allCollections: 'Todas las colecciones',
 
+  recentlyTranslatedHeading: 'Traducidas recientemente',
+  recentlyTranslatedSubtitle: 'Las obras más recientes que Source Library ha traducido a una versión moderna y legible.',
   discoverHeading: 'Descubre',
   discoverSubtitle: 'Fuentes primarias traducidas de la colección.',
   discoverEmpty: 'Explora la colección para descubrir fuentes primarias traducidas.',
